A problem was reported when quoted strings in command lines were being upper-cased.  This is not normal behaviour as DCL should respect the quotes. What is looks like is that DCL was parsing the command line, because it was defined via SET COMMAND.  However, it was then being parsed a second time by CLI$DCL_PARSE.
To fix this I have implemented a test at start up that checks to see if the
value of the DCL command line parser reserved word $VERB is in fact MMK.
If it is, then we assume that DCL has parsed the command line and all
qualifier values, etc. are set up.  If this value does NOT match, then
we fetch the command line with LIB$GET_FOREIGN and parse it as usual.

I have performed quite a bit of testing on this and it does seem to work.

This references #58.