Fix architecture built-in macro for x86_64
Rather than adding a fourth hard-coded option to the list, let's
derive the built-in name from the architecture name we already
have.  Some of them come back from sys$getsyi in mixed case, so
upper case the result when defining the built-in.

N.B.  The existing behavior differs from MMS and is not changed
here.  MMS will set, for example, MMSALPHA to a value of "ALPHA"
whereas MMK sets it to "1".  Either should work in an IFDEF, so
it may not matter.